fix: ensure bundledRaffleTicketsPerAdmission is treated as an integer
- Cast bundledRaffleTicketsPerAdmission to int for proper comparison in bundledRaffleId validation.

+++ includes/class-satoshi-api-client.php
@@ -411,7 +411,7 @@ final class SatoshiApiClient
                     $out[$k] = max(0, min(20, (int) $v));
                 } elseif ($k === 'bundledRaffleId') {
                     if ($v === null || $v === '') {
-                        if (($data['bundledRaffleTicketsPerAdmission'] ?? 0) === 0) {
+                        if ((int) ($data['bundledRaffleTicketsPerAdmission'] ?? 0) === 0) {
                             $out[$k] = null;
                         }
                     } else {
